API: move table to cogent3.core
