+++ changelog.d/20250406_125550_Gavin.Huttley.md
@@ -62,4 +62,6 @@ Uncomment the section that is right (remove the HTML comment wrapper).
   - `update()`
   - `union()`
 - `cogent3.format.alignment` has been renamed `cogent3.format.sequence`.
-   This is a backwards incompatible change!
+  This is a backwards incompatible change!
+- `cogent3.util.table` has been moved to `cogent3.core.table`. Usage of the
+  top level functions for making, loading tables is unaffected.
