+++ doc/api/table/classes/cogent3.util.table.Columns.rst
@@ -1,7 +1,7 @@
 Columns
 ======= 
 
-.. currentmodule:: cogent3.util.table
+.. currentmodule:: cogent3.core.table
 
 .. autoclass:: Columns
     :members:
