+++ src/cogent3/evolve/distance.py
@@ -5,10 +5,10 @@ from itertools import combinations
 from warnings import warn
 
 from cogent3 import make_tree
+from cogent3.core import table
 from cogent3.evolve.fast_distance import DistanceMatrix
 from cogent3.maths.stats.number import NumberCounter
 from cogent3.util import progress_display as UI
-from cogent3.util import table
 
 _NEW_TYPE = "COGENT3_NEW_TYPE" in os.environ
