+++ src/cogent3/parse/cogent3_json.py
@@ -2,7 +2,7 @@
 
 import json
 
-from cogent3.app.data_store import load_record_from_json
+from cogent3.app import data_store
 from cogent3.util.deserialise import deserialise_object
 from cogent3.util.io import open_
 from cogent3.util.misc import get_object_provenance
@@ -25,7 +25,7 @@ def load_from_json(filename, classes):
     with open_(filename) as f:
         content = json.loads(f.read())
     try:
-        _, data, completed = load_record_from_json(content)
+        _, data, completed = data_store.load_record_from_json(content)
         if not completed:
             msg = "json file is a record for type NotCompleted."
             raise TypeError(msg)
