+++ tests/test_format/test_bedgraph.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from cogent3.util.table import Table
+from cogent3.core.table import Table
 
 
 class FormatBedgraph(TestCase):
